requirements.txt
pytest==8.3.2
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
jsonschema==4.23.0
pytest-html==4.1.1
